+++ haedal_project/src/CardNews.js
@@ -1,70 +1,153 @@
-import "./CardNews.css";
-import "./embla.css";
-import React from "react";
-
-//npm i react react-dom @trendyol-js/react-carousel --save
-import { useEffect } from "react";
+import React, { useCallback, useEffect, useRef, useState } from "react";
 import useEmblaCarousel from "embla-carousel-react";
-import EmblaCarousel from "./EmblaCarousel";
+import imageByIndex from "./imageByIndex";
+import Autoplay from "embla-carousel-autoplay";
 
-export default function CardNews() {
-  const autoplayOptions = {
-    delay: 1000,
-    rootNode: (emblaRoot) => emblaRoot.parentElement,
-  };
+const mockApiCall = (minWait, maxWait, callback) => {
+  const min = Math.ceil(minWait);
+  const max = Math.floor(maxWait);
+  const wait = Math.floor(Math.random() * (max - min + 1)) + min;
+  setTimeout(callback, wait);
+};
 
-  const OPTIONS = {
-    inViewThreshold: 0,
-    dragFree: true,
-    containScroll: "keepSnaps",
-  };
-  const SLIDE_COUNT = 5; //슬라이드 개수 infinite
-  const SLIDES = Array.from(Array(SLIDE_COUNT).keys());
+const CardNews = (props) => {
+  const { options, slides: propSlides } = props;
+  const scrollListener = useRef(() => undefined);
+  const [slides, setSlides] = useState(propSlides);
+  const [emblaRef, emblaApi] = useEmblaCarousel(options, [Autoplay()]);
+  const [hasMoreToLoad, setHasMoreToLoad] = useState(true);
+  const [loadingMore, setLoadingMore] = useState(false);
+  const [pointerIsDown, setPointerIsDown] = useState(false);
 
-  const App = () => (
-    <main className="sandbox">
-      <section className="sandbox__carousel">
-        <EmblaCarousel slides={SLIDES} options={OPTIONS} />
-        {/*슬라이드 모션 추가*/}
-        {/*<div className="embla" ref={emblaRef} />*/}
-      </section>
-    </main>
-  );
+  const setPointerDown = useCallback(() => setPointerIsDown(true), []);
+  const setPointerNotDown = useCallback(() => setPointerIsDown(false), []);
 
-  return <App />;
-}
-/*
-    <div className="embla" ref={emblaRef}>
-      <div className="embla__container">
-        <img src={"/img/card1.png"} className="embla__slide" alt="" />
-        <img src={"/img/card2.png"} className="embla__slide" alt="" />
-        <img src={"/img/card3.png"} className="embla__slide" alt="" />
-        <img src={"/img/card4.png"} className="embla__slide" alt="" />
-        <img src={"/img/card5.png"} className="embla__slide" alt="" />
-        <img src={"/img/card1.png"} className="embla__slide_clone" alt="" />
-        <img src={"/img/card2.png"} className="embla__slide_clone" alt="" />
+  const lastSlideIsInView = useCallback(() => {
+    if (!emblaApi) return false;
+    const lastSlide = emblaApi.slideNodes().length - 1;
+    return emblaApi.slidesInView().indexOf(lastSlide) !== -1;
+  }, [emblaApi]);
+
+  const onScroll = useCallback(() => {
+    if (!emblaApi) return;
+    setLoadingMore((isLoadingMore) => {
+      if (isLoadingMore) return true;
+      const shouldLoadMore = lastSlideIsInView();
+      if (shouldLoadMore) emblaApi.off("scroll", scrollListener.current);
+      return shouldLoadMore;
+    });
+  }, [emblaApi, setLoadingMore, lastSlideIsInView]);
+
+  const addScrollListener = useCallback(() => {
+    if (!emblaApi || !hasMoreToLoad) return;
+    scrollListener.current = () => onScroll();
+    emblaApi.on("scroll", scrollListener.current);
+  }, [emblaApi, hasMoreToLoad, onScroll]);
+
+  const reloadEmbla = useCallback(() => {
+    if (!emblaApi) return;
+    const oldEngine = emblaApi.internalEngine();
+    emblaApi.reInit();
+    const newEngine = emblaApi.internalEngine();
+    Object.assign(newEngine.scrollBody, oldEngine.scrollBody);
+    Object.assign(newEngine.location, oldEngine.location);
+    Object.assign(newEngine.target, oldEngine.target);
+    const { index } = newEngine.scrollTarget.byDistance(0, false);
+    newEngine.index.set(index);
+    newEngine.animation.start();
+    setLoadingMore(false);
+  }, [emblaApi]);
+
+  useEffect(() => {
+    if (!emblaApi || slides.length === emblaApi.slideNodes().length - 1) return;
+    const engine = emblaApi.internalEngine();
+    const boundsActive = engine.limit.reachedMax(engine.target.get());
+    engine.scrollBounds.toggleActive(boundsActive);
+  }, [emblaApi, slides]);
+
+  useEffect(() => {
+    if (!emblaApi || !hasMoreToLoad || pointerIsDown) return;
+    if (slides.length === emblaApi.slideNodes().length - 1) return;
+    reloadEmbla();
+    addScrollListener();
+  }, [
+    emblaApi,
+    slides,
+    pointerIsDown,
+    hasMoreToLoad,
+    reloadEmbla,
+    addScrollListener,
+  ]);
+
+  useEffect(() => {
+    if (!emblaApi || hasMoreToLoad) return;
+    if (slides.length === emblaApi.slideNodes().length) return;
+    if (pointerIsDown && !lastSlideIsInView()) return;
+    reloadEmbla();
+    emblaApi.off("pointerDown", setPointerDown);
+    emblaApi.off("pointerUp", setPointerNotDown);
+  }, [
+    emblaApi,
+    slides,
+    hasMoreToLoad,
+    pointerIsDown,
+    setPointerDown,
+    setPointerNotDown,
+    reloadEmbla,
+    lastSlideIsInView,
+  ]);
+
+  useEffect(() => {
+    if (!emblaApi) return;
+    emblaApi.on("pointerDown", setPointerDown);
+    emblaApi.on("pointerUp", setPointerNotDown);
+    addScrollListener();
+  }, [emblaApi, setPointerDown, setPointerNotDown, addScrollListener]);
+
+  useEffect(() => {
+    if (!loadingMore) return;
+    mockApiCall(1000, 2000, () => {
+      setSlides((currentSlides) => {
+        if (currentSlides.length === 20) {
+          setHasMoreToLoad(false);
+          return currentSlides;
+        }
+        const newSlideCount = currentSlides.length + 5;
+        return Array.from(Array(newSlideCount).keys());
+      });
+    });
+  }, [setSlides, loadingMore]);
+
+  return (
+    <div className="embla">
+      <div className="embla__viewport" ref={emblaRef}>
+        <div className="embla__container">
+          {slides &&
+            slides.map((index) => (
+              <div className="embla__slide" key={index}>
+                <div className="embla__slide__number">
+                  <span>{index + 1}</span>
+                </div>
+                <img
+                  className="embla__slide__img"
+                  src={imageByIndex(index)}
+                  alt="Your alt text"
+                />
+              </div>
+            ))}
+          {hasMoreToLoad && (
+            <div
+              className={"embla-infinite-scroll".concat(
+                loadingMore ? " embla-infinite-scroll--loading-more" : ""
+              )}
+            >
+              <span className="embla-infinite-scroll__spinner" />
+            </div>
+          )}
+        </div>
       </div>
     </div>
-  );*/
-
-/*
-        <img src={"/img/카드들.png"} className="embla__slide" alt="" />
-        <img
-          src={"/img/카드들2.png"}
-          className="embla__container_clone"
-          alt=""
-        />
-        */
-
-/*
-        <img src={"/img/카드들.png"} className="embla__slide" />
-        <img src={"/img/카드들2.png"} className="embla__slide" />
-export default function CardNews() {
-  return ReactDOM.render(
-    <Carousel>
-      <img src={"/img/카드들.png"} className="image-wrap" />
-      <img src={"/img/카드들2.png"} className="image-wrap" />
-    </Carousel>,
-    document.getElementById("root")
   );
-}*/
+};
+
+export default CardNews;
