+++ haedal_project/src/Header.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+const Header = () => (
+  <header>
+    <h1 className="sandbox__header">Embla Carousel Autoplay React</h1>
+  </header>
+);
+
+export default Header;
