+++ haedal_project/src/sandbox.css
@@ -1,0 +1,46 @@
+.sandbox {
+    max-width: 67rem;
+    margin-left: auto;
+    margin-right: auto;
+  }
+  .sandbox__carousel {
+    background-color: var(--background-code);
+  }
+  .sandbox__carousel {
+    position: relative;
+  }
+  @media (min-width: 750px) {
+    .sandbox__carousel {
+      border-radius: 0.4rem;
+    }
+  }
+  .sandbox__header {
+    font-size: 2rem;
+    font-weight: bold;
+    text-align: center;
+    padding: 4rem 2rem 2rem 2rem;
+  }
+  .sandbox__footer {
+    display: flex;
+    justify-content: center;
+    padding: 2rem 2rem 4rem 2rem;
+  }
+  .sandbox__footer__link {
+    display: flex;
+    align-items: center;
+    background-color: transparent;
+    text-decoration: none;
+    border-radius: 3rem;
+    text-align: center;
+    font-weight: bold;
+    color: var(--text-low-contrast);
+    font-size: 1.4rem;
+  }
+  .sandbox__footer__link__svg {
+    display: block;
+    width: 2rem;
+    height: 2rem;
+    margin-right: 1rem;
+    fill: currentColor;
+  }
+
